Southfield Day 1 changes
Added logic to CollectorUp to move the arm if it needs to.
Added logic to Set the current arm position to Custom when the angle is
set manually.
Made low gear button also put the collector up (now TurtleStance)

Set the DriveTrain Expiration to 0.5 sec instead of 0.1 seconds,
disabled it during auton.
Properly initialized collector motors to victors